fix: canvas_danmaku 0.3.1 duration 参数类型改 double

+++ simple_live_app/lib/modules/live_room/player/player_controls.dart
@@ -653,7 +653,7 @@ Widget buildDanmuView(VideoState videoState, LiveRoomController controller) {
     option: DanmakuOption(
       fontSize: AppSettingsController.instance.danmuSize.value,
       area: AppSettingsController.instance.danmuArea.value,
-      duration: AppSettingsController.instance.danmuSpeed.value.toInt(),
+      duration: AppSettingsController.instance.danmuSpeed.value.toDouble(),
       opacity: AppSettingsController.instance.danmuOpacity.value,
       //strokeWidth: AppSettingsController.instance.danmuStrokeWidth.value,
       fontWeight: AppSettingsController.instance.danmuFontWeight.value,
